requirements.txt
discord.py>=2.0.0
yt-dlp>=2023.0.0
youtube-search-python>=1.6.0
lyricsgenius>=3.0.0
asyncio
logging